Improve site appearance
Moves the design of the site towards what might be considered acceptable
to a user.

- Display the task list as columns of cards.
- Move the task creation form into a card in that list.
- Adds a footer with a link to the project on github.
- Uses frontawesome and bootstrap.

In the future I intend to have all todo cards appear on initial loading
of the site, with the creation form redirecting to a index get request,
or not moving at all but refreshing the list through ajax or something.
I also intend to use fontawesome to change the form card into a button
card that pops up a lightbox with the form to let the body to be larger