replace site css with mobile friendly version
The common advice in css development is to develop for mobile first, and
add larger device styling on top of that. This is the commit that makes the
change to mobile styling, I plan to add media breaks later.

Now using css grid to organize section widths and heights, I prefer this
because all the sizing info is in one place rather than infered from
widths and hights of various disparate pieces.