Allow new tasks to be viewed without refreshing
Summarily: to improve user experience, and reduce network use.
When a user is accessing a webapp they expect the page they are viewing
to be a reflection of the current state of the system, this is not
accurate when the page needs to be refreshed to make changes.
With the prior method of refreshing after each task is added it puts the
onus on the user to request an update, which should not be needed when
they just requested to publish an update.
With the prior method, all the data was passed through the system after
the user refreshed the screen. Now that the user doesn't need to
refresh, one task goes through at a time, decreasing network use.

The website now uses an ajax post request to the server instead of a
form submission post request, the server responds with the id of the
task created, which the client sends back to the server in an ajax post
request to retrieve the html representation.
It is done this way to avoid replecating the html representation in the
javascript to submit this task, as that would break an architectural
boundary.
The resultant HTML is then added to the task container html element to
display on the page.

Future improvements could include: Displaying a pending task element to
show the page has responded to the player and isn't broken, or insert
the task right away with a pre-polled template and a later indicator
that it has been saved to db correctly; css animation to more smoothly
insert the task.