Fix an asnyc bug in MongoTaskGateway
Prior to this commit, there was a bug caught in the integration tests,
whereby access to the collection from the db connection was lost before
the opperation could do any work with it. This was not caught by unit
tests because all the database methods were stubbed out, and executed in
0 time.

wait for the opperation to execute before disconnecting from the
database.