Standardize architectual request response format.
Before this commit, the modules weren't communicating exclusively in
json as i had planned. This commit standardises this.

It also introduces much similarity between endpoints in the Router.
My goal is to eventually make the routing paths to the Controller completely
general and loaded from configuration information. To do this, it would be
far easier if each method followed the same control flow.